Enable OpenShift reporting and limited email by day
* Abstract common code for AWS and OpenShift reports
* Create new reporting for OpenShift
* Create schedule limit for emails less often
