+++ costemailer/rbac.py
@@ -8,6 +8,8 @@ ACCESS_ENDPOINT = "access/"
 
 AWS_ACCOUNT_ACCESS = "cost-management:aws.account:read"
 AWS_ORG_ACCESS = "cost-management:aws.organizational_unit:read"
+OPENSHIFT_CLUSTER_ACCESS = "cost-management:openshift.cluster:read"
+OPENSHIFT_PROJECT_ACCESS = "cost-management:openshift.project:read"
 
 
 def get_rbac_data(path="status/", params={}):
